migrated procedure in GraphRefactoring
