+++ core/src/main/java/apoc/nodes/Nodes.java
@@ -66,71 +66,6 @@ public class Nodes {
         }
     }
 
-    @Procedure(mode = Mode.WRITE)
-    @Description("apoc.nodes.deleteAndReconnect([pathLinkedList], [nodesToRemove], {config}) - Removes some nodes from a linked list")
-    public Stream<GraphResult> deleteAndReconnect(@Name("path") Path path, @Name("nodes") List<Node> nodesToRemove, @Name(value = "config", defaultValue = "{}") Map<String,Object> config) {
-
-        DeleteAndReconnectConfig conf = new DeleteAndReconnectConfig(config);
-
-        List<Node> nodes = IterableUtils.toList(path.nodes());
-        List<Relationship> rels = IterableUtils.toList(path.relationships());
-
-        if (!nodes.containsAll(nodesToRemove)) {
-            throw new RuntimeException("Some nodes in nodesToRemove not present in the path");
-        }
-
-        BiFunction<Node, Direction, Relationship> filterRel = (node, direction) -> StreamSupport
-                .stream(node.getRelationships(direction).spliterator(), false)
-                .filter(rels::contains)
-                .findFirst()
-                .orElse(null);
-
-        nodesToRemove.forEach(node -> {
-
-            Relationship relationshipIn = filterRel.apply(node, Direction.INCOMING);
-            Relationship relationshipOut = filterRel.apply(node, Direction.OUTGOING);
-
-            // if terminal node
-            if (relationshipIn == null || relationshipOut == null) {
-                rels.remove(relationshipIn == null ? relationshipOut : relationshipIn);
-
-            } else {
-                Node nodeStart = relationshipIn.getStartNode();
-                Node nodeEnd = relationshipOut.getEndNode();
-
-                RelationshipType newRelType;
-                Map<String, Object> newRelProps = new HashMap<>();
-
-                switch (conf.getRelationshipSelectionStrategy()) {
-                    case START:
-                        newRelType = relationshipIn.getType();
-                        newRelProps.putAll(relationshipIn.getAllProperties());
-                        break;
-                    case END:
-                        newRelType = relationshipOut.getType();
-                        newRelProps.putAll(relationshipOut.getAllProperties());
-                        break;
-                    default:
-                        newRelType = RelationshipType.withName(relationshipIn.getType() + "_" + relationshipOut.getType());
-                        newRelProps.putAll(relationshipOut.getAllProperties());
-                        newRelProps.putAll(relationshipIn.getAllProperties());
-                }
-
-                Relationship relCreated = nodeStart.createRelationshipTo(nodeEnd, newRelType);
-                newRelProps.forEach(relCreated::setProperty);
-
-                rels.add(relCreated);
-                rels.removeAll(List.of(relationshipIn, relationshipOut));
-            }
-
-            tx.execute("WITH $node as n DETACH DELETE n", Map.of("node", node));
-
-            nodes.remove(node);
-        });
-
-        return Stream.of(new GraphResult(nodes, rels));
-    }
-
     @Procedure
     @Description("apoc.nodes.get(node|nodes|id|[ids]) - quickly returns all nodes with these ids")
     public Stream<NodeResult> get(@Name("nodes") Object ids) {
