--- docs/asciidoc/modules/ROOT/examples/generated-documentation/apoc.nodes.deleteAndReconnect-lite.csv
@@ -1,2 +1,0 @@
-¦signature
-¦apoc.nodes.deleteAndReconnect(path :: PATH?, nodes :: LIST? OF NODE?, config = {} :: MAP?) :: (nodes :: LIST? OF NODE?, relationships :: LIST? OF RELATIONSHIP?)
