--- docs/asciidoc/modules/ROOT/pages/overview/apoc.nodes/apoc.nodes.deleteAndReconnect.adoc
@@ -1,44 +1,0 @@
-////
-This file is generated by DocsTest, so don't change it!
-////
-
-= apoc.nodes.deleteAndReconnect
-:description: This section contains reference documentation for the apoc.nodes.deleteAndReconnect procedure.
-
-label:procedure[] label:apoc-core[]
-
-[.emphasis]
-apoc.nodes.deleteAndReconnect([pathLinkedList], [nodesToRemove], \{config}) - Removes some nodes from a linked list
-
-== Signature
-
-[source]
-----
-apoc.nodes.deleteAndReconnect(path :: PATH?, nodes :: LIST? OF NODE?, config = {} :: MAP?) :: (nodes :: LIST? OF NODE?, relationships :: LIST? OF RELATIONSHIP?)
-----
-
-== Input parameters
-[.procedures, opts=header]
-|===
-| Name | Type | Default 
-|path|PATH?|null
-|nodes|LIST? OF NODE?|null
-|config|MAP?|{}
-|===
-
-== Config parameters
-include::partial$usage/config/apoc.nodes.deleteAndReconnect.adoc[]
-
-== Output parameters
-[.procedures, opts=header]
-|===
-| Name | Type 
-|nodes|LIST? OF NODE?
-|relationships|LIST? OF RELATIONSHIP?
-|===
-
-== Usage Examples
-include::partial$usage/apoc.nodes.deleteAndReconnect.adoc[]
-
-xref::graph-querying/node-querying.adoc[More documentation of apoc.nodes.deleteAndReconnect,role=more information]
-
