+++ docs/asciidoc/modules/ROOT/pages/overview/apoc.refactor/apoc.refactor.deleteAndReconnect.adoc
@@ -1,0 +1,43 @@
+////
+This file is generated by DocsTest, so don't change it!
+////
+
+= apoc.refactor.deleteAndReconnect
+:description: This section contains reference documentation for the apoc.refactor.deleteAndReconnect procedure.
+
+label:procedure[] label:apoc-core[]
+
+[.emphasis]
+apoc.refactor.deleteAndReconnect([pathLinkedList], [nodesToRemove], \{config}) - Removes some nodes from a linked list
+
+== Signature
+
+[source]
+----
+apoc.refactor.deleteAndReconnect(path :: PATH?, nodes :: LIST? OF NODE?, config = {} :: MAP?) :: (nodes :: LIST? OF NODE?, relationships :: LIST? OF RELATIONSHIP?)
+----
+
+== Input parameters
+[.procedures, opts=header]
+|===
+| Name | Type | Default 
+|path|PATH?|null
+|nodes|LIST? OF NODE?|null
+|config|MAP?|{}
+|===
+
+== Config parameters
+include::partial$usage/config/apoc.refactor.deleteAndReconnect.adoc[]
+
+== Output parameters
+[.procedures, opts=header]
+|===
+| Name | Type 
+|nodes|LIST? OF NODE?
+|relationships|LIST? OF RELATIONSHIP?
+|===
+
+[[usage-apoc.refactor.deleteAndReconnect]]
+== Usage Examples
+include::partial$usage/apoc.refactor.deleteAndReconnect.adoc[]
+
