+++ docs/asciidoc/modules/ROOT/partials/generated-documentation/nav.adoc
@@ -414,7 +414,6 @@ This file is generated by DocsTest, so don't change it!
 ** xref::overview/apoc.nodes/index.adoc[]
 *** xref::overview/apoc.nodes/apoc.nodes.collapse.adoc[]
 *** xref::overview/apoc.nodes/apoc.nodes.delete.adoc[]
-*** xref::overview/apoc.nodes/apoc.nodes.deleteAndReconnect.adoc[]
 *** xref::overview/apoc.nodes/apoc.nodes.get.adoc[]
 *** xref::overview/apoc.nodes/apoc.nodes.group.adoc[]
 *** xref::overview/apoc.nodes/apoc.nodes.link.adoc[]
@@ -463,6 +462,7 @@ This file is generated by DocsTest, so don't change it!
 *** xref::overview/apoc.refactor/apoc.refactor.cloneSubgraph.adoc[]
 *** xref::overview/apoc.refactor/apoc.refactor.cloneSubgraphFromPaths.adoc[]
 *** xref::overview/apoc.refactor/apoc.refactor.collapseNode.adoc[]
+*** xref::overview/apoc.refactor/apoc.refactor.deleteAndReconnect.adoc[]
 *** xref::overview/apoc.refactor/apoc.refactor.extractNode.adoc[]
 *** xref::overview/apoc.refactor/apoc.refactor.from.adoc[]
 *** xref::overview/apoc.refactor/apoc.refactor.invert.adoc[]
