--- docs/asciidoc/modules/ROOT/partials/usage/config/apoc.nodes.deleteAndReconnect.adoc
@@ -1,10 +1,0 @@
-The procedure support the following config parameters:
-
-.Config parameters
-[opts=header, cols="1,1,1,5"]
-|===
-| name | type | default | description
-| relationshipSelectionStrategy | Enum[START, END, MERGE] | true | if START, the incoming relationship will be attached to the next node, if END the outcoming relationship will be attached, if MERGE will be attacched a merge of start and end relatioships. With the same property name, takes precedence the start node.
-| relTypesToAttach | List<String> | emptyList | If not empty, it represents the list of relation types that will attach to the remaining nodes. It must be of the same size as the list of nodes to be eliminated. Takes precedence over the `attachStartRel` (if not empty)
-| relsToAttach | List<Relationship> | emptyList | If not empty, it represents the list of relations that will attach to the remaining nodes. It must be of the same size as the list of nodes to be eliminated. Takes precedence over the `attachStartRel` and `relTypesToAttach` (if not empty)
-|===
